Add request output into nested tabs
